pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "claudemonitor"
version = "0.1.0"
description = "Local dashboard + native window for monitoring Claude Code sessions via HTTP hooks"
readme = "install.md"
requires-python = ">=3.9"
dependencies = [
    "pywebview>=5.0",
]

[tool.setuptools]
py-modules = ["server", "launcher"]
